Adds dynamic learning for hammering repetitions
We start by hammering each pattern once but at PROBES_PER_PATTERN
different locations. If we found a bit flip, we do a reproducibility check
(do random accesses, then repeat hammering the same pattern again at
the same location; for 100 times). We derive from this check the number
of repetitions we need on average to trigger a bit flip. We use the current
value of repetitions and compute the sliding/running average of the current
and the new value. Using this, we learn how often we need to hammer a pattern
to see any bit flips. Ideally, all the time we can keep 1 repetition
(e.g., on Samsung).

Signed-off-by: Patrick Jattke <pjattke@ethz.ch>