Adds test_location_dependence
This method checks whether moving only the aggressor pair
that triggered the bit flip, rather than moving all
aggressor pairs during the sweep, leads to more bit flips.
In that case, we consider aggressors in the pattern to
be location-dependent, i.e., relying on the absolute
address. We then fix then when doing the full sweep.

Signed-off-by: Patrick Jattke <pjattke@ethz.ch>