fix: issues in post-stage of fuzzing run
Method sweep_pattern modified the original mapping but did not restore it afterward.
Now this method creates a copy of the original mapping and restores it before returning,
so that subsequent sweeps will start from the same position.

We had to limit the number of patterns that we use during the reproducibility check as
computing the reproducibility score for all pattern takes too long.

A threshold was added in the method find_direct_effective_aggs that finds the
AggressorAccessPattern that caused a bit flip. If the distance is larger than the
threshold, we don't consider it as being the one who caused the bit flip anymore. This
could be, for example, when remapping is going on.

Removed leftover debugging printouts.

Signed-off-by: Patrick Jattke <pjattke@ethz.ch>